requirements.txt

# ── Core API ──────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-multipart==0.0.9

# ── Auth ──────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Groq (LLM + Embeddings) ───────────────
groq==0.9.0

# ── Qdrant (Vector DB) ────────────────────
qdrant-client==1.9.1

# ── MongoDB (Student Memory) ──────────────
motor==3.4.0
pymongo[srv]==4.7.2

# ── Hybrid Search ─────────────────────────
rank-bm25==0.2.2

# ── Sarvam AI (Voice) ─────────────────────
requests==2.32.3
aiohttp==3.9.5
pydub==0.25.1

# ── Multilingual ──────────────────────────
langdetect==1.0.9

# ── Text Processing ───────────────────────
regex==2024.4.28
numpy==1.26.4
tiktoken==0.7.0

# ── Async ─────────────────────────────────
aiofiles==23.2.1
httpx==0.27.0

# ── Logging ───────────────────────────────
structlog==24.2.0

# ── Testing ───────────────────────────────
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0

bcrypt==4.1.3